refactor: consolidate foxy deposit code
